functional, preliminary ble binary data transmission

+++ Hardware/Firmware/mainboard/include/config.h
@@ -25,12 +25,12 @@
 #define THRESH_NO_TOUCH 30
 
 struct reading {
-    float ax;
-    float ay;
-    float az;
-    float gx;
-    float gy;
-    float gz;
+    int16_t ax;
+    int16_t ay;
+    int16_t az;
+    int16_t gx;
+    int16_t gy;
+    int16_t gz;
 };
 
 struct offsets {
